+++ packages/theme/docs/.vitepress/works.ts
@@ -66,8 +66,8 @@ const workConfig: Theme.UserWorks = {
       time: {
         start: '2020/02/18'
       },
-      github: 'https://github.com/ATQQ/sugar-blog',
-      url: 'https://sugarat.top',
+      github: 'https://github.com/cnjimbo/cnjimbo.github.io',
+      url: 'https://www.dmsrs.org',
       tags: ['Vitepress', 'Vue'],
       cover:
             'https://img.cdn.sugarat.top/mdImg/MTY3MzE3MDUxOTMwMw==673170519303',