+++ packages/vitepress-plugin-pagefind/CHANGELOG.md
@@ -139,4 +139,4 @@
 
 ### Minor Changes
 
-- feat: 从 [@sugarat/theme](https://github.com/ATQQ/sugar-blog/tree/master/packages/theme) 主题分离独立的 pagefind 插件
+- feat: 从 [@sugarat/theme](https://github.com/cnjimbo/cnjimbo.github.io/tree/master/packages/theme) 主题分离独立的 pagefind 插件